Refactored classes, reduce complexity, renamed files, removed hardcoded toke
