+++ 3582/1/top_gems.rb
@@ -1,17 +1,14 @@
 # frozen_string_literal: true
 
 require 'optparse'
-require 'yaml'
-require './factory'
 require 'pry'
-require 'terminal-table'
+require './gem_factory'
+require './table_output'
+require './gem_reader'
 
-# Main class, it parses commands, reads file, fetches info about gems,
-# outputs concrete (specified) gems
+# Main class
 class TopGems
-  HEADING_TABLE = ['name', 'used by', 'watchers', 'stars', 'forks', 'contributors', 'issues'].freeze
-
-  attr_reader :params, :gems, :machina
+  attr_reader :params, :gems
 
   def initialize
     @params = {}
@@ -28,27 +25,6 @@ class TopGems
 
   private
 
-  def print_gems
-    table_str = @gems.map do |gemi|
-      create_string_table(gemi)
-    end
-    print_table(table_str)
-  end
-
-  def create_string_table(gemi)
-    [
-      gemi.gem_name, gemi.params[:used_by],
-      gemi.params[:watchers], gemi.params[:stars],
-      gemi.params[:forks], gemi.params[:contributors],
-      gemi.params[:issues]
-    ]
-  end
-
-  def print_table(tb_data)
-    table = Terminal::Table.new headings: HEADING_TABLE, rows: tb_data
-    puts table
-  end
-
   def reduce_gems(top, name)
     @gems = @gems.slice(0..(top - 1)) if top.is_a?(Integer) && top <= @gems.size && top.positive?
     @gems.select! { |elm| elm.gem_name.include?(name) } if name.is_a?(String)
@@ -58,8 +34,12 @@ class TopGems
     @gems.sort! { |first, second| first.rank < second.rank ? 1 : -1 }
   end
 
+  def print_gems
+    TableOutput.new(@gems).execute
+  end
+
   def fetch_gems
-    gem_names = open_file(params[:file])
+    gem_names = GemReader.new(params[:file]).execute
     gem_names.each do |name|
       @gems << GemFactory.build(name)
     end
@@ -73,18 +53,6 @@ class TopGems
     end.parse!(into: @params)
 
     @params[:top] = @params[:top].to_i if @params[:top]
-  end
-
-  def open_file(file)
-    begin
-      raise Errno::ENOENT unless file.is_a?(String)
-
-      gem_names = YAML.load_file(file).dig('gems')
-    rescue Errno::ENOENT
-      gem_names = %w[sinatra rspec rails nokogiri]
-      puts "File Doesn't exist or not specifeid, loading deafult gems"
-    end
-    gem_names
   end
 end
